@@ -3,6 +3,6 @@
 	Copyright (c) 2023 - 2025 TNO-ESI
 	All rights reserved.
 """
-__version__ = '0.2.7'
+__version__ = '0.2.14'
 
 from .base import MBDElement, MBDNet, MBDNode, MBDReasoner, MBDRelation, MBDNetReasonerView, longest_common_fqn
