@@ -4,6 +4,7 @@
 	All rights reserved.
 """
 import re
+from multiprocessing import Queue
 from typing import Union, Optional
 
 from capellambse import MelodyModel
@@ -11,8 +12,8 @@ from capellambse.metamodel import cs, fa
 
 from neomodel import db
 
-from mbdlyb.functional.gdb import (Cluster, DiagnosticTest, DiagnosticTestResult, Function, Hardware, OperatingMode,
-								   RequiredForRelation, DirectObservable, update_fqn)
+from mbdlyb.functional.gdb import (Cluster, DiagnosticTest, DiagnosticTestResult, Function, Hardware, ObservedError,
+								   OperatingMode, RequiredForRelation, DirectObservable, update_fqn)
 
 
 class CapellaToCluster:
@@ -38,12 +39,26 @@ class CapellaToCluster:
 	_diagnostic_test_mapping: dict[DiagnosticTestResult, list[str]]
 	_function_positions: dict[str, tuple[Cluster, Hardware]]
 
-	def __init__(self, aird: str) -> None:
+	_capella_fns: dict[str, fa.Function]
+	_error_exchanges: set[str]
+	_initial_error_propagations: dict[str, dict[str, list[str]]]
+	_non_self_error_reporting_functions: set[Function]
+	_error_code_nodes: dict[str, ObservedError]
+
+	_queue: Optional[Queue]
+
+	def __init__(self, aird: str, queue: Queue = None) -> None:
 		self._aird = aird
 		self._model = self._load_model()
 		self._arch_dict = {"logical": self._model.la, "physical": self._model.pa}
 		self._diagnostic_test_mapping = dict()
 		self._function_positions = dict()
+		self._capella_fns = dict()
+		self._error_exchanges = set()
+		self._initial_error_propagations = dict()
+		self._non_self_error_reporting_functions = set()
+		self._error_code_nodes = dict()
+		self._queue = queue
 
 	def transform(
 			self,
@@ -52,7 +67,7 @@ class CapellaToCluster:
 			default_hl_fn_tests: bool = False,
 			default_hw_tests: bool = False,
 			default_ll_hw_tests: bool = False,
-			import_opms: bool = False,
+			import_opms: bool = False
 	) -> Cluster:
 		"""Transforms Capella architecture to Neomodel Cluster."""
 		self._set_architecture(architecture)
@@ -62,17 +77,30 @@ class CapellaToCluster:
 		self._default_ll_hw_tests = default_ll_hw_tests
 		self._import_opms = import_opms
 
+		self._update_status(0.)
+
 		self._cluster = self._build_component_tree(self._architecture.root_component)
+		self._update_status(.1)
 		self._build_functional_tree(self._architecture.root_function)
+		self._update_status(.2)
 
 		self._connect_test_results()
+		self._update_status(.3)
 		self._position_remaining_nodes()
+		self._update_status(.4)
 		update_fqn(self._cluster)
+		self._update_status(.5)
 
+		self._add_error_propagation_relations()
+		self._update_status(.6)
 		self._add_required_for_relations()
+		self._update_status(.7)
+		self._add_observed_errors()
+		self._update_status(.7)
 
 		if self._import_opms:
 			self._add_operating_modes()
+		self._update_status(.8)
 
 		# add default tests
 		if self._default_fn_tests:
@@ -90,10 +118,16 @@ class CapellaToCluster:
 			for n in self._elements[Hardware].values():
 				if not any(f.has_subfunctions for f in n.realizes.all()):
 					self._add_default_diagnostic_test(n)
+		self._update_status(.9)
 
 		# ensure unique FQNs
 		self._format_fqns()
+		self._update_status(1.)
 		return self._cluster
+
+	def _update_status(self, status: float):
+		if self._queue is not None:
+			self._queue.put_nowait(status)
 
 	def _build_component_tree(self, capella_comp: cs.Component) -> Optional[Cluster]:
 		if not capella_comp.components and not capella_comp.allocated_functions:
@@ -110,25 +144,51 @@ class CapellaToCluster:
 				subcluster.set_net(c)
 		return c
 
-	def _build_functional_tree(self, capella_fn: fa.Function) -> Optional[Function]:
-		if self._fn_root_of_diagnostic_tree(capella_fn):
-			if self._fn_is_diagnostic_test(capella_fn):
+	def _build_functional_tree(self, capella_fn: fa.Function, parent_dt: DiagnosticTest = None)\
+				-> Optional[Function]:
+		name = self._format(capella_fn.name)
+		fn: Function = None
+		if self._fn_is_diagnostic_test(capella_fn):
+			if parent_dt is None:  # we have a top-level test
 				cost = capella_fn.property_value_groups['MBDlyb.Diagnostic test']['Cost']
-				_, dtr = self._add_specified_diagnostic_test(self._format(capella_fn.name), cost)
-				self._diagnostic_test_mapping[dtr] = [f.uuid for f in capella_fn.functions]
-			for child in capella_fn.functions:
-				self._build_functional_tree(child)
-		else:
+				dt, dtr = self._add_specified_diagnostic_test(self._format(capella_fn.name), cost)
+				self._diagnostic_test_mapping[dtr] = [f.uuid for f in
+													  list(capella_fn.functions) + [e.source.owner for i in
+																					capella_fn.inputs for e in
+																					i.exchanges]]
+				for child in capella_fn.functions:
+					self._build_functional_tree(child, dt)
+				return None
+			else:  # we have a subtest, we ignore the cost
+				fn = self._add_function(capella_fn)
+				dtr = DiagnosticTestResult(name=name + '_Result').save()
+				parent_dt.test_results.connect(dtr)
+				self._diagnostic_test_mapping[dtr] = [capella_fn.uuid]  # for now
+		elif parent_dt is not None:  # subfunction of a test
 			fn = self._add_function(capella_fn)
-			children = [self._build_functional_tree(child_fn) for child_fn in capella_fn.functions]
-			for subfunction in children:
-				if subfunction is not None:
-					subfunction.subfunction_of.connect(fn)
+			# self._add_direct_observables_to_fn(capella_fn, fn)  # maybe later
+		elif self._fn_root_of_diagnostic_tree(capella_fn):  # no function to be added here
+			for child in capella_fn.functions:
+				self._build_functional_tree(child, parent_dt)
+			return None
+		else:  # normal function
+			fn = self._add_function(capella_fn)
 			self._add_direct_observables_to_fn(capella_fn, fn)
-			return fn
-		return None
 
-	def _add_hardware(self, capella_hw: cs.Component, cluster: Cluster) -> Hardware:
+		children = [self._build_functional_tree(child_fn, parent_dt) for child_fn in
+					capella_fn.functions]
+		for subfunction in children:
+			if subfunction is not None:
+				subfunction.subfunction_of.connect(fn)
+		return fn
+
+	def _add_hardware(self, capella_hw: cs.Component, cluster: Cluster) -> Optional[Hardware]:
+		complete_composite = False
+		if capella_hw.property_value_groups.get('MBDlyb.Composite'):
+			complete_composite = capella_hw.property_value_groups['MBDlyb.Composite']['Complete composite']
+		if complete_composite:
+			return None
+
 		fault_rate = 0.01  # default
 		if capella_hw.property_value_groups.get('MBDlyb.Hardware priors'):
 			fault_rate = capella_hw.property_value_groups['MBDlyb.Hardware priors']['Fault rate']
@@ -143,12 +203,14 @@ class CapellaToCluster:
 		return hw
 
 	def _add_function(self, capella_fn: fa.Function) -> Function:
+		self._capella_fns[capella_fn.uuid] = capella_fn
 		fn = Function(name=self._format(capella_fn.name)).save()
 		self._add_node(capella_fn, fn)
 		if capella_fn.uuid in self._function_positions:
 			c, hw = self._function_positions[capella_fn.uuid]
 			fn.set_net(c)
-			fn.realized_by.connect(hw)
+			if hw is not None:
+				fn.realized_by.connect(hw)
 		return fn
 
 	def _add_specified_diagnostic_test(self, name: str, cost: float, cluster: Cluster = None) -> tuple[
@@ -162,13 +224,74 @@ class CapellaToCluster:
 		return dt, dtr
 
 	def _add_direct_observables_to_fn(self, capella_fn: fa.Function, fn: Function):
-		for uo in [o for o in capella_fn.outputs if not o.exchanges]:
-			if bool(uo.applied_property_value_groups) and uo.property_value_groups['MBDlyb.Direct observables'][
-				'Observable']:
-				do = DirectObservable(name=self._format(uo.name)).save()
-				do.observed_functions.connect(fn)
-				if capella_fn.uuid in self._function_positions:
-					do.set_net(self._function_positions[capella_fn.uuid][0])
+		for o in [_o for _o in capella_fn.outputs if _o.applied_property_value_groups]:
+			do = None
+			if 'MBDlyb.Error reporting' in o.property_value_groups.keys():
+				is_reporting = o.property_value_groups['MBDlyb.Error reporting']['Reporting errors']
+				ec = o.property_value_groups['MBDlyb.Error reporting']['Error code']
+				if is_reporting and bool(ec):
+					do = ObservedError(name=self._format(o.name), error_code=ec).save()
+					do.observed_function.connect(fn, {'error_code': ec})
+				elif is_reporting:
+					self._non_self_error_reporting_functions.add(fn)
+				elif bool(ec):
+					for ex in o.exchanges:
+						self._error_exchanges.add(ex.uuid)
+						if capella_fn.uuid not in self._initial_error_propagations:
+							self._initial_error_propagations[capella_fn.uuid] = {ex.target.parent.uuid: [ec]}
+						else:
+							if ex.target.parent in self._initial_error_propagations[capella_fn.uuid]:
+								self._initial_error_propagations[capella_fn.uuid][ex.target.parent.uuid].append(ec)
+							else:
+								self._initial_error_propagations[capella_fn.uuid][ex.target.parent.uuid] = [ec]
+
+			if not o.exchanges:
+				if 'MBDlyb.Direct observables' in o.property_value_groups.keys() and \
+						o.property_value_groups['MBDlyb.Direct observables']['Observable']:
+					do = DirectObservable(name=self._format(o.name)).save()
+					do.observed_functions.connect(fn)
+			if do is not None and capella_fn.uuid in self._function_positions:
+				do.set_net(self._function_positions[capella_fn.uuid][0])
+
+	def _add_error_propagation_relations(self):
+		for source_uuid, target_data in self._initial_error_propagations.items():
+			source_fn = self._elements[Function].get(source_uuid)
+			for target_uuid, ecs in target_data.items():
+				for ec in ecs:
+					oe = ObservedError(name=self._format(ec), error_code=ec).save()
+					oe.observed_function.connect(source_fn, {'error_code': ec})
+					self._error_code_nodes[ec] = oe
+				target_fn = self._elements[Function].get(target_uuid)
+				source_fn.communicates_through.connect(target_fn, {'error_codes': ecs})
+				self._propagate_error_relations(self._capella_fns[target_uuid], ecs)
+
+	def _propagate_error_relations(self, cfn: fa.Function, ecs: list[str]):
+		source_fn = self._elements[Function][cfn.uuid]
+		for o in cfn.outputs:
+			if o.exchanges:
+				for e in o.exchanges:
+					target = e.target.parent
+					target_fn = self._elements[Function][target.uuid]
+					if source_fn.required_for.is_connected(target_fn):
+						rel = source_fn.required_for.relationship(target_fn)
+						rel.error_codes += ecs
+						rel.save()
+					else:
+						source_fn.required_for.connect(target_fn, {'error_codes': ecs})
+						self._error_exchanges.add(e.uuid)
+					self._propagate_error_relations(target, ecs)
+
+	def _add_observed_errors(self):
+		for of in self._non_self_error_reporting_functions:
+			for predecessor in of.requires.all():
+				for rel in of.requires.all_relationships(predecessor):
+					for ec in rel.error_codes:
+						# do = ObservedError(name=self._format(ec), error_code=ec).save()
+						do = self._error_code_nodes[ec]
+						do.reporting_function.connect(of, {'error_code': ec})
+						if of.get_net() is not None:
+							do.set_net(of.get_net())
+
 
 	def _load_model(self) -> MelodyModel:
 		"""Loads Capella model using capellambse library."""
@@ -292,8 +415,11 @@ class CapellaToCluster:
 		"""
 		capella_fn_exchanges = []
 		for exch in self._architecture.all_function_exchanges:
-			src_uuid = exch.source.owner.uuid
-			trgt_uuid = exch.target.owner.uuid
+			if exch.uuid in self._error_exchanges:
+				continue
+
+			src_uuid = exch.source.parent.uuid
+			trgt_uuid = exch.target.parent.uuid
 
 			if (
 					src_uuid in self._elements[Function]
@@ -311,7 +437,7 @@ class CapellaToCluster:
 			self._elements[RequiredForRelation][exch_uuid] = rel
 
 	def _position_remaining_nodes(self) -> None:
-		# Position unpositioned functions
+		# Position unpositioned functions based on their subfunctions
 		db.cypher_query('''CALL () {
     MATCH (root:Cluster)
     WHERE NOT (root)-[:PART_OF]->(:Cluster)
@@ -322,10 +448,30 @@ class CapellaToCluster:
 WITH *
 UNWIND dislocated_functions AS f
 CALL (f, root) {
-    MATCH (sf:Function)-[:SUBFUNCTION_OF]->(f)
-    WITH collect(sf) AS sfs
-    MATCH (c:Cluster)
-    WHERE ALL (sf IN sfs WHERE (sf)-[:PART_OF*]->(c))
+    MATCH (sf:Function)-[:SUBFUNCTION_OF*]->(f)
+    MATCH (c:Cluster)<-[:PART_OF]-(sf)
+    WITH collect(c) as cs, collect(sf) as sfs
+    MATCH (cr:Cluster)
+    WHERE size(cs) > 0 AND all(cx IN cs WHERE (cx)-[:PART_OF*]->(cr))
+    MATCH p = (cr)-[:PART_OF*]->(root)
+    RETURN cr, cs, p, length(p) AS length
+    ORDER BY length(p) DESC
+    LIMIT 1
+}
+CREATE (f)-[:PART_OF]->(cr)''')
+		# Position unpositioned functions based on their superfunctions
+		db.cypher_query('''CALL () {
+    MATCH (root:Cluster)
+    WHERE NOT (root)-[:PART_OF]->(:Cluster)
+    MATCH (f:Function)
+    WHERE NOT (f)-[:PART_OF]->(:Cluster)
+    RETURN root, collect(f) AS dislocated_functions
+}
+WITH *
+UNWIND dislocated_functions AS f
+CALL (f, root) {
+    MATCH (sf:Function)<-[:SUBFUNCTION_OF*]-(f)
+    MATCH (c:Cluster)<-[:PART_OF]-(sf)
     MATCH p = (c)-[:PART_OF*]->(root)
     RETURN c, p, length(p) AS length
     ORDER BY length(p) DESC
