@@ -7,8 +7,12 @@ from typing import Optional
 
 from neomodel import DoesNotExist
 from nicegui import ui
+from itertools import product
+from math import prod
 
-from mbdlyb.gdb import MBDElement
+from quimb.tensor.tensor_2d_tebd import conditioner
+
+from mbdlyb.gdb import MBDElement, MBDNode
 
 
 def goto(url: str):
@@ -77,9 +81,115 @@ class Relation:
 		}
 
 
+class CPTCell:
+	line: 'CPTLine'
+	probability: float
+
+	def __init__(self, line: 'CPTLine', probability: float):
+		self.line = line
+		self.probability = probability
+
+	@property
+	def cpt(self) -> 'CPT':
+		return self.line.cpt
+
+
+class CPTLine:
+	cpt: 'CPT'
+	conditions: list[str]
+	cells: list[CPTCell]
+
+	def __init__(self, cpt: 'CPT', conditions: list[str], probabilities: list[float] = None):
+		self.cpt = cpt
+		self.conditions = conditions
+		self.cells = [CPTCell(self, f) for f in (probabilities or ([0.] * len(self.values)))]
+
+	def __repr__(self):
+		return f'{", ".join(self.conditions)} - {" ".join(str(c.probability) for c in self.cells)}'
+
+	@property
+	def values(self) -> list[str]:
+		return self.cpt.values
+
+	def normalize(self) -> 'CPTLine':
+		flatten = False
+		total = sum(c.probability for c in self.cells)
+		if total == 0.:
+			flatten = True
+			total = len(self.values)
+		for c in self.cells:
+			c.probability = (1. if flatten else c.probability) / total
+		return self
+
+
+class CPT:
+	parents: list[MBDNode]
+	node: MBDNode
+	lines: list[CPTLine]
+
+	def __init__(self, node: MBDNode, lines: list[CPTLine] = None):
+		self.node = node
+		self.lines = self.create_lines(lines)
+
+	def __repr__(self):
+		return 'CPT for {}\n{} - {}\n{}'.format(self.node, ", ".join(p.name for p in self.parents),
+												", ".join(self.values), "\n".join(str(l) for l in self.lines))
+
+	@property
+	def parents(self) -> list[MBDNode]:
+		return self.node.parents()
+
+	@property
+	def values(self) -> list[str]:
+		return self.node.states
+
+	def create_lines(self, _lines: list[list[float]] = None) -> list[CPTLine]:
+		lines: list[CPTLine] = []
+		condition_list = list(list(condition) for condition in product(*[p.states for p in self.parents]))
+		if _lines is None:
+			_lines = [[0.] * len(self.values) for _ in condition_list]
+		for idx, conditions in enumerate(condition_list):
+			lines.append(CPTLine(self, conditions, _lines[idx]))
+		return lines
+
+	def normalize(self) -> 'CPT':
+		for line in self.lines:
+			line.normalize()
+		return self
+
+	def set_prob(self, state: str, prob: float):
+		state_idx = self.values.index(state)
+		for line in self.lines:
+			line.cells[state_idx].probability = prob
+
+	def to_dict(self) -> dict:
+		return {
+			'state_count': len(self.values),
+			'parents': [[p.fqn, len(p.states)] for p in self.parents],
+			'lines': [[c.probability for c in l.cells] for l in self.lines]
+		}
+
+	@staticmethod
+	def from_dict(node: MBDNode, d: dict) -> tuple['CPT', Optional[str]]:
+		message = None
+		if not d:
+			message = 'No CPT found.'
+		elif not all(p in d for p in ('state_count', 'parents', 'lines')):
+			message = 'Invalid CPT provided.'
+		elif d['state_count'] != len(node.states):
+			message = 'Number of states has changed.'
+		elif d['parents'] != [[p.fqn, len(p.states)] for p in node.parents()]:
+			message = 'Dependencies have changed.'
+		elif len(d['lines']) != prod(len(p.states) for p in node.parents()):
+			message = 'Number of stored CPT lines is invalid.'
+		if message is not None:
+			return CPT(node), message
+		return CPT(node, d['lines']), None
+
+
 # DATABASE HELPERS
 
-def get_object_or_404(cls, **kwargs):
+def get_object_or_404[T](cls: type[T], **kwargs) -> T:
 	try:
 		return cls.nodes.get(**kwargs)
 	except DoesNotExist:
